+++ event-protocol/ruby/lib/cucumber/events/version.rb
@@ -1,0 +1,5 @@
+module Cucumber
+  module Events
+    VERSION = "0.1.0"
+  end
+end
