--- event-protocol/ruby/lib/cucumber/formatter/event_stream.rb
@@ -1,14 +1,0 @@
-require "cucumber/formatter/event_stream/version"
-require "cucumber/formatter/event_stream/plugin"
-
-module Cucumber
-  module Formatter
-    module EventStream
-
-      def self.new(*args)
-        Plugin.new(*args)
-      end
-
-    end
-  end
-end
