--- event-protocol/ruby/lib/cucumber/formatter/event_stream/version.rb
@@ -1,7 +1,0 @@
-module Cucumber
-  module Formatter
-    module EventStream
-      VERSION = "0.1.0"
-    end
-  end
-end
